fastapi/requirements.txt
authlib
fastapi
geoalchemy2
httpx
itsdangerous
psycopg2-binary
pydantic
pyjwt
python-dotenv
requests
sqlalchemy
uvicorn